+++ coloraide/color/distance/__init__.py
@@ -1,45 +1,24 @@
 """Distance and Delta E."""
-from ... import util
 from . import distance_euclidean
-from . import delta_e_76
-from . import delta_e_94
-from . import delta_e_cmc
-from . import delta_e_2000
-from . import delta_e_itp
-from . import delta_e_99o
-from . import delta_e_jz
-from . import delta_e_hyab
-
-SUPPORTED = [
-    delta_e_76, delta_e_94, delta_e_cmc, delta_e_2000,
-    delta_e_itp, delta_e_99o, delta_e_jz, delta_e_hyab
-]
+from abc import ABCMeta, abstractmethod
 
 
-class DeltaEMap(util.Map):
-    """
-    Immutable delta E mapping.
+class DeltaE(ABCMeta):
+    """Delta E plugin class."""
 
-    This is not required to be used by users but is
-    more for internal use to discourage people from
-    accidentally altering the base class mapping.
-    """
+    @staticmethod
+    @abstractmethod
+    def name():
+        """Get name of method."""
 
-    def __init__(self, values):
-        """Delta E mapping."""
-
-        self._d = {value.__name__.split('.')[-1].replace('delta_e_', ''): value.distance for value in values}
-
-    def _error(self):  # pragma: no cover
-        """Error message."""
-
-        return util.ERR_MAP_MSG.format(name="DE_MAP")
+    @staticmethod
+    @abstractmethod
+    def distance(color, sample, **kwargs):
+        """Get distance between color and sample."""
 
 
 class Distance:
     """Distance."""
-
-    DE_MAP = DeltaEMap(SUPPORTED)
 
     def delta_e(self, color, *, method=None, **kwargs):
         """Delta E distance."""